backend/requirement.txt
flask==3.1.0
flask-cors==6.0.0
Pillow==10.4.0
torch==2.7.0
torchvision==0.22.0